refactor(character-extraction): split step into components/ directory
CharacterExtractionStep slimmed to root branching only. Extracted:
- components/RefImageTray.tsx   — tray + thumbnail with signed URL fetching
- components/CharacterCard.tsx  — resting card
- components/CharacterModal.tsx — self-contained modal, reads hook directly
- components/CharacterList.tsx  — list container + blur backdrop + activeId state

CharacterModal now calls useCharacterExtraction/useSceneEngine directly,
eliminating prop drilling through CharacterList.
